requirements.txt
torch~=2.2.0
transformers~=4.37.2
google-generativeai~=0.3.2
numpy~=1.26.3
PyAudio~=0.2.14
pillow~=10.2.0
pygame~=2.5.2